+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Unified SMTP config (#257, PLUM Sprint 230825)
 - Endpoint for pruning old audit entries (#259, PLUM Sprint 230825)
 - Algorithmic (stateless) anonymous sessions (#232, PLUM Sprint 230825)
+- Initialize Zookeeper service (#265, PLUM Sprint 230908)
 
 ### Refactoring
 - Last login info moved to a dedicated collection (#259, PLUM Sprint 230825)