+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Endpoint for pruning old audit entries (#259, PLUM Sprint 230825)
 - Algorithmic (stateless) anonymous sessions (#232, PLUM Sprint 230825)
 - Initialize Zookeeper service (#265, PLUM Sprint 230908)
+- Initialize Sentry service (#262, PLUM Sprint 230908)
 
 ### Refactoring
 - Last login info moved to a dedicated collection (#259, PLUM Sprint 230825)